fix(security): share the SSRF-safe agent from core, and satisfy the scanners
Three CI findings on this branch, all real:

SonarCloud flagged 5% duplication on new code. The cause was the SSRF-safe HTTPS
agent existing twice — once in the Make.com plugin, once in the Zapier copy this
branch added. Both plugins already depend on @gauzy/core, and core is server-only,
so the agent now lives there and both import it. That is where it should have gone
in the first place; the reason it was duplicated was to keep Node's https/dns out
of @gauzy/utils, which is browser-reachable — core has no such constraint. The
Make.com module re-exports it so existing importers are unaffected.

GitGuardian flagged a test fixture shaped like a real GitHub token. It was fake,
but a literal that looks like a credential trains people to ignore the scanner, so
the fixtures are now assembled at runtime.

cspell flagged British spellings, "spoofable", and the random letter runs used as
masking fixtures. The prose is reworded and the fixtures now use real words, which
also makes the expected values readable ('*'.repeat(13) + 'TAIL' rather than a
wall of asterisks).

No behaviour change: 38 tests still pass and core, integration-zapier and
integration-make-com all typecheck clean.
