+++ packages/core/src/lib/core/util/index.ts
@@ -2,3 +2,4 @@ export * from './find-by-date-between';
 export * from './path-util';
 export * from './db-operators';
 export * from './organization-sensitive-relations.config';
+export * from './ssrf-safe-agent';
