+++ packages/plugins/integration-zapier/src/lib/webhook-url.validator.ts
@@ -1,8 +1,7 @@
-import * as https from 'https';
-import * as dns from 'dns';
-import type { LookupFunction } from 'net';
 import { BadRequestException } from '@nestjs/common';
-import { getUnsafeOutboundUrlReason, isPrivateOrLoopbackHost } from '@gauzy/utils';
+import { getUnsafeOutboundUrlReason } from '@gauzy/utils';
+
+export { createSsrfSafeHttpsAgent } from '@gauzy/core';
 
 /**
  * Validates a tenant-supplied Zapier webhook target URL against the SSRF egress guard.
@@ -13,6 +12,10 @@ import { getUnsafeOutboundUrlReason, isPrivateOrLoopbackHost } from '@gauzy/util
  * before each outbound delivery, so subscriptions stored before this guard existed are also
  * rejected (GHSA-6gg6-vv4f-2x74).
  *
+ * The connection-time half of the guard is `createSsrfSafeHttpsAgent` (re-exported above from
+ * `@gauzy/core`), which re-checks the resolved IP and so catches DNS rebinding that a URL-literal
+ * check cannot.
+ *
  * @param url - The webhook target URL to validate.
  */
 export function assertSafeZapierWebhookUrl(url: string): void {
@@ -20,40 +23,4 @@ export function assertSafeZapierWebhookUrl(url: string): void {
 	if (reason) {
 		throw new BadRequestException(`Invalid webhook URL: ${reason}`);
 	}
-}
-
-/**
- * Creates an HTTPS agent whose DNS resolver rejects any hostname that resolves to a private,
- * loopback or link-local address.
- *
- * Validating the RESOLVED IP at connection time (not just the URL literal) closes DNS-rebinding /
- * hostname-based SSRF: a public-looking hostname that resolves to an internal IP is refused before
- * the socket connects, with no time-of-check/time-of-use gap (GHSA-6gg6-vv4f-2x74).
- *
- * @returns An `https.Agent` that blocks connections to non-public addresses.
- */
-export function createSsrfSafeHttpsAgent(): https.Agent {
-	const lookup: LookupFunction = (hostname, options, callback) => {
-		dns.lookup(
-			hostname,
-			options as dns.LookupAllOptions,
-			(err: NodeJS.ErrnoException | null, address: any, family: any) => {
-				if (err) {
-					return (callback as any)(err, address, family);
-				}
-				// `address` is a string when `options.all` is falsy, or an array of { address, family } otherwise.
-				const candidates: string[] = Array.isArray(address) ? address.map((entry) => entry.address) : [address];
-				const blocked = candidates.find((ip) => isPrivateOrLoopbackHost(ip));
-				if (blocked) {
-					const error = Object.assign(
-						new Error(`SSRF blocked: ${hostname} resolved to a non-public address (${blocked})`),
-						{ code: 'ESSRFBLOCKED' }
-					);
-					return (callback as any)(error, address, family);
-				}
-				return (callback as any)(null, address, family);
-			}
-		);
-	};
-	return new https.Agent({ lookup });
 }
